rag_agent/requirements.txt
google-adk
google-cloud-aiplatform
google-cloud-storage
pandas
openpyxl
xlrd
requests
python-dotenv
